IDEA-253911 Wait for scopes to load in Find Usages tests
Now that Find Usages uses ScopeService to load scopes,
it's no longer enough to call
NonBlockingReadActionImpl.waitForAsyncTaskCompletion
to ensure that Find Usages was successfully completed.

Add a similar method to ScopeService to wait for it.
Introduce a higher level method waitForAsyncTaskCompletion
into FindUsagesManager so the tests don't have to deal
with implementation details.

GitOrigin-RevId: 972fcf2c7eaaaf1fa7cef4a3263c0e3c04772c1e