[gradle][groovy] IDEA-257715 fix a bug when delegate = T after resolving.
GradleDelegatesToProvider was wrongly determining a delegate for a closure in `Project#configure` method. The method has three overloads:
```
Object configure(Object object, Closure configureClosure);

Iterable<?> configure(Iterable<?> objects, Closure configureClosure);

<T> Iterable<T> configure(Iterable<T> objects, Action<? super T> configureAction);
```
When any of the first two versions are called, GradleDelegatesToProvider searches for a method having Action and tries to get a delegate from there. It founds the third method implementation, but is not able to resolve the T type parameter. So it returns just T as a delegate. This has no sense.

GitOrigin-RevId: 70a3001ee0c8584c4328143fd1ac8cdf685f5986