[gradle][groovy] IDEA-257715 add comment
Probably this code needs to be reworked. Currently, to define a delegate in a Closure, it searches for an overloaded method with Action to get a delegate from its generic. It doesn't verify a whole signature, just checks if the last parameter is Action.

It works well for cases like these two Project methods:
```
CopySpec copySpec(@DelegatesTo(CopySpec.class) Closure closure);

CopySpec copySpec(Action<? super CopySpec> action);
```

Let's imagine a class where method signatures are more different:
```
interface FooClass {
    void method(String name, Closure configure) // is called
    void method(String name, Action<? super Dependency> configure)
    String method(Action<? super Project> configure)
}
The first two versions of the method have a similar signature, most probably it would be okay to get a delegate from the second method with Action<Dependency>. But the signature of the third method looks different - probably it has a slightly different purpose. If this method is processed earlier than the second, a delegate could be taken from there. Would it be correct?

If there are any similar cases in Gradle classes, they could have potential issues with delegate resolve.

GitOrigin-RevId: 38893f2efb7fe13fa55d857644b679f22a4c1e77