[module-based loader] fix locating product content modules when the embedded frontend process is started from full IDE (RDCT-1226)
The frontend part of the IDE has its own list of content modules, and at least one of them ('intellij.cwm.guest.notebooks') is not included in lib/modules directory, and it's not simple to put it there, because it's currently part of CWM plugin distribution. So 'findProductContentModuleClassesRoot' is introduced in 'ProductLoadingStrategy', which computes path to the classes root of the module using data from the module repository if the module-based loader is used.

GitOrigin-RevId: dddcb8b16dce64fd73a0cbdde507daaa5b069a10