remove exact matches without template text by index check
In RemDev, on frontend, we do `(merge (customize (merge (front back))) customized-on-backend (back))` children. Due to the merge, `ActionUrl.absolutePosition` can be incorrect. But we can customize backend actions on frontend. That would not be possible if we did `(merge (customize (front)) customized-on-backend (back))`.

Fixes
GTW-7803 Broken customized menu in Split Mode: Services in Tools case

GitOrigin-RevId: 9ef937a98324f60287392077fd107c88fa04b86c