IDEA-249914 Extract display info troubleshooting collector code
We'll need this in two places at least: troubleshooting.txt
and submitting bug reports from the IDE. Getting all the numbers
right is tricky, so we better have one place for it.

And while we're at it, add insets to troubleshooting.txt.
They're sometimes very useful when debugging sizing and positioning
issues.

GitOrigin-RevId: 728ad646391b136ad963c7f9f80a361285cb15b6