IDEA-253911 Wait for the scopes to load before calculating find usage options
The scope is picked from the dialog, regardless of whether the dialog
was actually shown or not. However, if it isn't shown, then the scopes
aren't loaded yet, and therefore fetching the selected scope only works
for predefined scopes, because they're pre-loaded synchronously
in the scope combo constructor.

To make it work for all scopes, wait until the combo initializes
and only then close the dialog (it isn't really shown, so it's just
about disposing it), compute the settings and proceed to find usages.

The tricky part is that we have to do it under modal progress because
otherwise we risk the PsiElement (the one we're searching for)
being invalidated. It actually happens in some tests like
TypeScriptFindUsagesTest (when they're all run sequentially)
and may happen in production as well, throwing an assertion error.

GitOrigin-RevId: c8b6ccab3125539fd17e32ba088d04429e387d01