API dump: don't include synthetic bridges in the dump
These methods exist to support
- covariant return type overrides;
- generic overrides;
- overrides of `void` methods with methods returning `java/lang/Void` in Kotlin.

Such methods are called only through supertype invocations, and supertypes are already listed in the dump.

The main reason for avoiding these in dumps: Java allows extending a public class `B` from a package-local class `A`.
`B` can override a method of `A` with a covariant return type.
In this case, a synthetic bridge is generated in `B` with the return type from method in `A`.
If the method return type in `A` is also a package-local class `C`, this would mean `C` is leaked by the API surface of `B`.
Instead, we consider such bridges not a part of the API. For a concrete example, see:
`com.intellij.tools.apiDump.testData.inheritFromPackageLocal.JPackageLocalClass.packagePrivateMethodWithPackagePrivateReturnType`.

Another reason has the same root cause: Java allows such class inheritance.
A public class `B` which extends package-local class `A` "opens" up _all_ public/protected methods of `A` to the outer world.
In subsequent commits, this is going to be reflected in the dump by copying the API methods of package-local `A` into public `B`.
We don't want to copy the bridges, which a part of `A` and can be only accessed through supertype of `A`,
to avoid false positive leaking API errors.

GitOrigin-RevId: 2aecf62ad18f4ed6458927ccbd86f8066bea4751