IDEA-260517 Expand All in Project View to affect only selected node
Add a registry key to hide the regular Expand All action
and replace it with a new action that only expands
the selected subtree.

We already have an action for this, but it's bound to NumPad*
by default, which is awkward to use on TKL keyboards and smaller,
and it also seems broken for async tress, the Project View included.
So instead add a new action that's tailored to the Project View.
Reuse the Expand shortcut from the editor for this action,
which is the same as Expand All.

Add another action to mimic the usual Expand All behavior,
but bind it to Expand All editor shortcut.

Add isExpandAllEnabled to TreeExpander to be able to distinguish
between
"we can expand all, but don't want the Expand All action to work"
(canExpand() == true, isExpandEnabled() == false)
and
"we can't expand all"
(canExpand() == false).

GitOrigin-RevId: 40bd24953e360a265fc4695b9df0119ff1e039af