[gutters] Correctly scale left free painters area with icons (IDEA-343272)
#IDEA-343272 Fixed

This only attacks the issue for the LEFT painters area, the RIGHT painters area still needs to be looked at.

Icons are scaled proportional to JBUI.scale(_) as they're loaded
because of scaling done by `CachedImageIcon`.

Therefor, the free painters area that is expected to contain (mostly) icons should be scaled by UI scaling.
There are two options for this:
1. We make the pixel width reserved unscaled
2. We make the pixel width reserved scaled by UI scaling

Preferably, we could do 1, as it means more consistently storing unscaled values and scaling is then applied directly on repaint.
However, this can be tricky because of how gutter sizes are calculated in various places.
Here the most challenging thing is usually how to adjust for editor font size.

GitOrigin-RevId: 668365f8c8c1a357ddbf316d4c15280f814cf3c3