[module-based loader] don't fail startup if product content module cannot be found (CPP-38280)
We use the same root JetBrainsClientPlugin.xml for all IDE-specific frontends, and not all product content modules specified in it are available in all IDEs. E.g., 'intellij.notebooks.ui' isn't present in CLion. Since content modules are optional, absence of some of the in the distribution shouldn't cause an error.

GitOrigin-RevId: 074f93d1a5c31de6d6bf5152d96ba150c72bf7b2