EA-1145457 Add NPE check in getClassIconOfKind
This function is only called in two places in PsiClassImplUtil.java,
and in both places it's expected to return a non-null value.
Since we have no way of figuring out why it returned null
(and it doesn't seem reasonably possible), let's try
to throw a more meaningful exception (since one is going
to be thrown anyway) and see what happens.

While we're at it, specify nullability for both input and output.

GitOrigin-RevId: 36523f24ab7218200d34deb82766bb580a8b6215