[debugger] IDEA-351117 Fix evaluation on thread incompatible with evaluation
After evaluation failed with IncompatibleThreadStateException exception, the debugger marks that thread as incompatible with evaluation and redirects all evaluations to the primary thread (which is likely in the good state). It will cover most of the rendering problems but may produce some confusion if the evaluating expression relays on some TheadLocal or Thread.currentThread(). So it would be nice to add some notification about another-thread evaluation state.

GitOrigin-RevId: b5d0df97c48ef8026da221a245cded12068d461d