[debugger] IDEA-350602 Remove assertion: suspend-thread context can point to another thread while stepping in it
Suppose, two threads (A and B) in suspend-thread mode are suspended and stepping performed in one of them (A). To avoid context switch, the debugger sets active execution stack in the context of the other suspended thread (B) to the thead performed stepping (A). So the removed assertion was not correct in this situation.

GitOrigin-RevId: 684f73090982ceee6adc17227aa85326a4946805