IJPL-426 rework IndexingProgressReporter.{CheckCancelOnlyProgressIndicator > CheckPauseOnlyProgressIndicator}
checkCanceled should always be done using "current context" (thread-local progress indicator or coroutine context), not with some outer progress indicators (outer progress indicators should be installed into "current context" before invocation)

This change might break the behavior of scanning in *dumb* mode. We don't care about it too much, because this mode will be removed soon anyway: IDEA-325937. As of now, we only need it to work fine in unit tests to ease migration to indexing in the background thread

GitOrigin-RevId: 5e74cd156b43829f5ec96c4688f93496fcd80aa0