IDEA-346842 Align gradient to the device pixel grid
Yet another fractional scaling mess.

First, align the Graphics painting rectangle origin to the device grid
(alignTxToInt). This guarantees that x=0 is located at some device
pixel exactly.

Then align all three coordinates: left x, center and right x
to the device grid using alignIntToInt. The choice of rounding
modes is more or less random, but since the gradient is small enough
already, we use FLOOR for the left coordinate and CEIL for the right one,
as it's better to err on the larger side. For the central coordinate,
use CEIL for similar reasons: the left part if often even smaller,
for example, when displaying the menu in a separate toolbar on Windows.

GitOrigin-RevId: 6e658e5158e2b34983dd48ef41b57ee4ed91a8ee