[AI Assistant] LLM-2457 Add the ability to toggle chat history in popup mode
The toggle option enables users to view and hide chat history when using chat in a popup. This includes modifications to the UI inputs and actions associated with the chat popup, and the addition of "Restore Last Chat" functionality. The change is guarded behind the "llm.chat.in.popup" registry flag to allow for safe rollout and usability feedback collection.

GitOrigin-RevId: 5d691a5349db333c273ebdebde2e31587c5eff1c