[build scripts] resolve included product-modules.xml file in source roots, not in outputs (IJPL-957)
To determine the list of bundled plugins in modular loader, we need to load the root product-modules.xml and resolve other product-modules.xml files included in it. Since the project may not be compiled at that moment (it's needed to know which plugins are bundled to decide which modules to compile), included product-modules.xml files must be loaded from sources, not from output roots (where RuntimeModuleRepository points to).

GitOrigin-RevId: b7b919e715aa0a1ff605ec70cf054f84a089ec53