Improve CircularProgressIndicator code (#352)
We only use the launched effect to load the new frames, and instead use
an infinitely repeating animation for the actual frame selection.
This is done so that UI tests don't hang on the infinite work done in
the launched effect. We also now allow providing a dispatcher on which
we should load the frames, which defaults to Default, but can be
overridden in tests with a TestDispatcher.
GitOrigin-RevId: d5ee158342c29cb051ba8cae11ffafdd64ad8270