GTW-7931 Preserve the order of options and give configurable's options higher priority than registrar's
In `ActionAsyncProvider` we merge options from the registrar with options gathered from configurables directly.

In some cases, the hits in these options are duplicated and only one has to be chosen.
But `MutableSet` doesn't guarantee order of the elements when we're iterating through it, so sometimes an option from registrar is chosen, sometime an option from configurables.

The most noticeable case is the choice between "Plugins" page (`preferences.pluginManager`) and a page created by a plugin itself.
E.g. for "Kotlin Notebook" plugin an option will lead to `kotlinNotebook` page, but for "Liquibase" plugin an option will lead to `preferences.pluginManager`

To have more predictable choice, we can use `LinkedHashSet` which preserves the order of the added elements.

We also want to ensure that configurables have a higher priority than registrar options, so they get chosen as a result option.
To ensure that we add them at the beginning of the `optionDescriptions` and after them, registrar's options are added.
Then during iteration they're going to be processed first.

GitOrigin-RevId: af41cfa7455101b449b3189e9bf75ded35f7073d