Add theme name to ThemeDefinition and JewelTheme (#353)
Names are supposed to be unique per each theme. In standalone, they're
hardcoded to "IntUI Light" and "IntUI Dark", and in the bridge they
match the Swing LaF name.

This allows us to fix a bug in Markdown in the bridge where switching
between two LaFs with the same isDark value would cause the Markdown
text not to recompose appropriately.
GitOrigin-RevId: 7872a6d2821eea5a66141711b22aec8be6d6786d