[terminal] IJPL-101428 Always reset prompt in Bash
But remember the original prompt value in `__JETBRAINS_INTELLIJ_ORIGINAL_PS1` variable and report it in `prompt_state_updated` event. So, there will be no situation when there is no prompt, because we reset it and no one updated it.
Also, `prompt_updated` event is now fired when generator is finished.

GitOrigin-RevId: 7cee7d4d871afd6a17362b28367fb6f05104ba68