GTW-8455 Fix frame is shown too early when creating a new project
In the commit below project session creation was moved from `InitProjectActivity` to `ProjectServiceContainerInitializedListener` in order to have the valid project session as early as possible (before any components or services are loaded)
https://jetbrains.team/p/ij/repositories/ultimate/revision/8918289b326ae6f5657ed5d8ad1976edb7dfd67d

However, in case of new project creation, `ProjectServiceContainerInitializedListener`s are executed before the project is actually opened.
In monolith (standalone IDE) it's not noticeable because the frame is created much later. But in split (RD) we create the frame at the same time, we receive new project info from backend.

I partially revert changes - `ProjectAttachActivity` becomes `InitProjectActivity` again, so we can be sure that frontend receives project information only when the project is actually opened.

GitOrigin-RevId: be21c03ced8bf0287ec7439ce2c50a2a8e2cdd8b