[terminal] IJPL-101428 Take into account line wraps when calculating cursor position in the prompt
The cursor position in the terminal is based on a position in a grid. But some lines can be wrapped there, so when we transform the grid to string, the number of lines can differ. Because the string consists of logical lines, where all wrapped lines are merged into one logical line. So, we need to transform the cursor position too.

GitOrigin-RevId: 8aa3778383608e2210d84d597879cc99332ab944