IJPL-1002 Implement saving presentation data in TreeState
To fix the Project View jumping up and down when
its state is loaded, we need to persist its presentation data,
as there's no way to load it quickly enough. Implement it
by adding new createOn() overloads to TreeState that
persist these presentations in addition to the expanded
and/or selected paths.

Because the state is restored incrementally, loading
nodes, then loading their children, and so on, we need
some way to merge cached presentations with those
that have been already loaded. To implement this,
we persist not just presentations, but also their
path IDs, the same as used for expanded/selected paths.

To match cached nodes to already loaded nodes,
we add the new getPathElementType() method
to the PathElementIdProvider interface and implement
it in the cached node class. Normally the type is calculated
based on the real class, but for cached nodes we need their
type to match the type of the real node.

The CachedTreePresentationSupport interface will be used
by consumers of this saved state, such as tree models and trees.

For now, persist only texts. We can add other attributes,
such as colors, later. Persisting icons is tricky, but likely not needed,
as "loading" spinning icons are a good way to tell which nodes
have been loaded already.

Implement a mutable structure to hold the entire tree as well.
It'll be needed in the tree and its model, and those are already
complicated enough, so let's keep most logic out of them.
This structure holds not just cached nodes, but also the real
ones as they're loaded by the model, so we can continue
displaying cached children for the already loaded nodes.
This structure is short-lived and only exists while the tree
is being loaded and expanded, so it's OK to cache anything
there, it'll be all cleared soon enough.

GitOrigin-RevId: 8b4bf5f5be0659fdbdae8ba5d5f52b6584b0c00a