[terminal] IJPL-101428 Fix TerminalDeletePreviousWordTest
It was broken because `TerminalDeletePreviousWordAction` required `TerminalPromptController` and it was not added to the context in the test.
Now the test is additionally configured to use TerminalPrompt language in the editor and contain both prompt and the command in the document. Also, added the test to check that this action is not removing the prompt text.

GitOrigin-RevId: c9b7962d188ae4e7edc8aec2582a39aec57a51a5