[gradle][groovy] IDEA-269370 add more strict conditions for files that could be a navigation target for plugin ID.
These conditions match some requirements for Precompiled script plugins. There are more strict conditions for Precompiled plugins, not considered here because they require more complicated implementation.

See more information about Precompiled plugins in the Gradle documentation:
https://docs.gradle.org/current/userguide/custom_plugins.html#sec:precompile_script_plugin">Gradle Precompiled plugins

According to the documentation, there is an opportunity to add a package declaration for a plugin implemented either on Groovy or Kotlin. In practice, I was able to create a plugin having a package only on Kotlin. That's why I did not consider a case when plugin ID contains packages in `findPrecompiledGroovyPlugin` method

GitOrigin-RevId: b273e40013c57c0b2b9fd7cf65c4b61b407a3dca