IJPL-148398 Fix invisible divider dragging
It was explicitly allowed to drag an invisible divider
if at least one of the components is showing.
It is not clear why anyone would need to drag the divider
if one of the components isn't showing, but to be on the safe
side, let's leave that behavior alone and instead disallow
dragging when one of the components is null or invisible
(as opposed to not showing). Invisible means that it was
explicitly hidden by code, so it's unlikely that we'll need
to drag the divider (as is demonstrated by the issue).

GitOrigin-RevId: 41ec6f2a9d1605b2d278edfedd4f3133b70684cf