[terminal] Make most of the New Terminal classes internal
Some classes are implementation detail or not intended to be used in the plugins. Some classes can be considered as an API, but better to make them internal for now, because New Terminal is under active development. We don't need an external contributions now and don't know what API will be needed, so better to not leak anything now.

GitOrigin-RevId: 42559c8b7ee9eb9ea143c89f6752a05e8bb8e2d9