IJPL-48505 Add ide.dnd.threshold registry key
Use it to override the OS default value for the DnD gesture
recognition threshold.

It's very tricky to properly initialize it: on one hand,
the registry must be already loaded, on the other hand,
some gesture recognizers (macOS) use static initialization
and therefore we should override the value before the relevant
class is loaded. Thankfully, the DnD manager seems to be
the perfect place. It's where drag gesture recognized are actually
created, and the registry seems to be already loaded at this point.

GitOrigin-RevId: b2bb169494d8bf99dfb7ad8e3709e9e604b9a0d8