IJPL-73517 IJent + WSL: don't rely on existence of `intellij.platform.ijent.impl` plugin
The plugin with IJent binaries and the client implementation is not available in the community repo. Therefore, it's not bundled with the IDEs built outside our build infrastructure.

However, the WSL support belongs to the community repo, so it should work in every IDE, with or without `intellij.platform.ijent.impl`.

Due to a logical bug, the IDE decided to use IJent for launching WSL processes if and only if the registry flag `wsl.use.remote.agent.for.launch.processes` is set. The flag was defined right in the core, and the flag is set to true by default. Hence, IDEs without the installed IJent plugin (AndroidStudio, DataGrip 241.14494) couldn't execute anything on WSL.

This commit moves the registry flag right into the implementation plugin. Now, an IDE doesn't try to use IJent for WSL if the plugin is not installed.

(cherry picked from commit 24c947740ab1baafc25a74f5794904a86e4f0c30)

IJ-MR-131941

GitOrigin-RevId: 089e61ae1fd425f62cbf1b15b087bcdf5561a810