Fix menu dividers in bridge (#361)
The main root cause was a missing return statement in retrieveIntAsDp(),
but we also had terribly inconsistent values for the menus compared to
the actual appearance. I have tweaked the bridge menu style defaults to
align better, but we'll have to come back to revisit the whole menu
implementation to actually use the right values in the right way. The
current implementation is not satisfactory.
GitOrigin-RevId: 5ffff1124420179750630370aaa33888f88506f2