IJPL-1002 Implement cached presentations in Tree / AsyncTreeModel
To use cached presentations saved/restored by TreeState,
we implement the new CachedTreePresentationSupport
interface in both Tree and AsyncTreeModel. The tree part
handles expand/collapse logic, the model part handles
the tree structure itself.

To keep cached nodes expanded as the mode loads real nodes,
we listen to tree structure change / node insertion events
and re-expand nodes accordingly. Failure to do so would lead
to nodes collapsing as soon as their real counterparts are loaded.
However, if a node is explicitly collapsed, for example,
because the user clicked it, we remember that and don't
re-expand these nodes again as they're re-added as children
of nodes being loaded.

GitOrigin-RevId: 11c85e5f38c9a9f63ce499328e2d408c53b77ff1