PY-71801 IJent: quick fix for sending SIGINT
It was forgotten to provide support for sending SIGINT to processes, which is important in many Python applications: SIGINT leads to the well-known KeyboardInterrupt error, while SIGTERM invokes usual signal handlers.

This commit brings a rough-and-ready implementation of SIGINT that strictly relies on the tool /bin/kill, which is usually available on WSL.

Later, the ability to send SIGINT to a process should be implemented in the IJent binary directly.

IJ-CR-132132

(cherry picked from commit 2579f0eda1b32a818d3c9fcb56f90532146baa1c)

GitOrigin-RevId: ca52032ec31826b2e3f1ae6631b3ec3afd97c81f