IJPL-1002 Implement icon saving/restoring for tree presentation cache
An icon is easily loaded if we know its main three coordinates:
the path, the plugin ID and the module ID. The tricky part is
to get these from an existing icon to save the coordinates
in the cache. Since we have a lot of different icon types,
and apparently no utility method to extract this from any icon,
the best we can do for now is to implement this for the most
common icon types, unwrapping them as we go. Seems to be
working for most icons for now.

GitOrigin-RevId: 76adae6f25ffd1cab18fbbebfbda205b8ce9e1a5