GTW-7926 Create `FrontendSpotlightPainter` in `ThinClientSettingsDialog` to update spotlights for host's settings
Add `updateSpotlights` method to `RemoteSettingsSession` model which sends the signal to the backend with the text filter which should be used to calculate spotlights.

The signal is sent by `FrontendSpotlightPainter` when `update` method is called.

The `update` method is called on filter change (using MergingUpdateQueue), focus lost/gain and configurable change.
The configurable for the update is calculated in `SettingsEditor` based on the current selected page in the tree view.

To reduce the amount of data sent from frontend to backend, we can send updates only when active configurable is host's configurable.

GitOrigin-RevId: e126d8313236f3c212cfb3ff6011fd8ef111179a