GTW-7926 Fix irrelevant spotlights are drawn for the selected page
`SpotlightPainter` subscribes on `highlightComponent` topic (ComponentHighlightingListener.TOPIC) which is updated by a static `highlightComponent` method in `SearchUtil`

All spotlights are sent through the same message bus and received by all painters.

Before, only one `SpotlightPainter` existed, so there was no need to check that the component belongs to the painter.
But now, in RD, we create a new painter per each page which causes the wrong spotlight to be drawn for the wrong pages.

To fix that, we can check that `myTarget` is an ancestor of the highlighted component and should be added to the glass panel to be drawn.
The same thing, for example, has been done for `CodeStyleAbstractPanel`.

GitOrigin-RevId: 5713621292a75ebf4110628ca37108f6997d3309