Add onTextClick to MarkdownBlockRenderer (#366)
* Fix text colour in Markdown (bridge)

We were not setting the right text colour when creating the bridge
Markdown Styling. Now we make sure we do respect the colour
set in the various InlinesStyle's TextStyles when rendering inlines.

In addition, I've added a simple API to be able to initialize a
MarkdownStyling in the bridge based only on a base TextStyle. This is
useful, e.g., when you want to make smaller text such as comments.

* Add onTextClick to MarkdownBlockRenderer

This is a workaround to the fact that ClickableText swallows all clicks,
since its onClick lambda has no way of saying if it did handle a click.

Hopefully this won't be needed anymore when we move to Compose 1.7,
which brings in a proper LinkAnnotation.
GitOrigin-RevId: 007b5d1c0def14d9888b631620058fdac5ea0b88