IJPL-1002 Save leaf states of cached Project View nodes
To show expand controls for those nodes that were collapsed
when the state was cached, we need to know whether the node
was a leaf or not. So save it in a new attribute. Be sure to fall
back to false if the node actually has cached children, just in case
XML was corrupted or something like that.

Return null instead of an empty list when the model requests
children for non-leaf nodes without cached children, so the
model will show a "Loading..." node instead of just an empty
list.

GitOrigin-RevId: 2f4c685656017c3f07fd9875df9daac581eede5f