IJPL-872 single `KernelService` implementation for all backends
This service, when loaded as part of the CWM plugin, tries to override `MonolithKernelService`.
We are avoiding service overrides with a proper module structure.

CWM runs as monolith+plugin, so `MonolithKernelService` is loaded,
but it does not install the workspace clock like `BackendKernelService` does.

Solution 0: reliable and correct.
It should be possible to register `MonolithKernelService` for the local client ID,
while CWM plugin would register `BackendKernelService` for remote clients.
In split backend mode, there should be no monolith registration at all,
and it should not be needed because all clients are remote.

Solution 1: okay.
A single service implementation must be used in both monolith and split host.
This way there will be no conflict between the monolith service and the CWM plugin service.
However, the workspace clock will also be installed in the monolith even if it might be unneeded,
e.g., when the CWM plugin is not loaded.

Solution 1 is used in this commit.

GitOrigin-RevId: bcb75ac0dabc3328bd09c30312510a93bbead11b