[terminal] IJPL-101589 Add shell name to the ShellRuntimeContext
ShellType class can't be used there because it is enum.
It is impossible to add the new value to the enum without source code modification. So, it means that if we allow shell integrations implementation in the external plugins, they will be unable to add their shell type to this enum.
Value class is a more suitable and extendable approach here.

GitOrigin-RevId: 48ad8aff21e2670bd84e656abc42accffd2c4735