IJent tests cleanup: delete ImaginaryFileTree and MockUnixIjentFileSystemApi
It's better to test IJent on real file systems. Unit tests in the IDEA Aggregator won't bring any benefits to IJent, because it's independent of the changes in various extension points, build scripts, etc.

Being in the aggregator was the only reason to have unit tests for such a specific feature.

GitOrigin-RevId: 0407005412ed9b3e60eebc22744fc4d21591c2a3