[tools] provide a main class to launch a product which uses the module-based loader from sources and compile the necessary modules
To run a product which uses the module-based loader using the same configuration of classloaders and plugins as in production (RDCT-1194), we need to have dependencies of IntellijLoader only in the original application classpath. This means that we cannot specify some aggregating *.main module in the run configuration, and therefore the IDE won't be able to compile the necessary modules before starting the run configuration. As a workaround, 'Build Project' can be specified as a 'Before launch' step, but compilation of the whole project may take a lot of time. This commit introduces IntellijDevLauncher, which uses BuildHttpRequestHandler to trigger compilation of necessary modules from the process.

GitOrigin-RevId: f355c32daba1989b97a8d3df7216cc2233f28579