IJPL-1002 Don't update new StructureTreeModel children twice
It's a very common case: when the children are loaded for the first time,
they're updated from Node's constructor. And then they're updated again
in getChildren(). Sometimes it does make sense to update them in getChildren()
because the model may reuse old child nodes if they didn't change, but for
freshly created children it doesn't make any sense.

Fix by keeping track of newly created children and then only updating
those that aren't newly created. To protect against concurrent and reentrant
calls, make the tracking set thread-local and reentrant by keeping a count
of reentrant calls. It's likely an unnecessary precaution, but better be safe
than sorry. Release the set by clearing the reference when we're done
to avoid leaks.

GitOrigin-RevId: c6247c68bbe47d7c19dde2190d6b0da5ce433ceb