IJPL-1002 Make "External Libraries" never-leaf
The default is to load all its children just to check whether it's a leaf.
This slows down Project View loading for no good reason,
especially if the node is never expanded.

The only downside of a never-leaf node is that it'll always show
an expand control, even if there are no actual children.
Which is a very small trade-off that only affects very exotic cases.

GitOrigin-RevId: 39a0924475dc5e63f02f8ab548601b66e115deeb