IJPL-1002 Make library nodes never-leaf
The default is to load all its children just to check whether it's a leaf.
This slows down Project View loading for no good reason,
especially if the node is never expanded, which is often the case
for a library node.

The only downside of a never-leaf node is that it'll always show
an expand control, even if there are no actual children.
Which is a very small trade-off that only affects very exotic cases
when some library exists but contains nothing.

GitOrigin-RevId: 4957320db7c26353bf26ff61bc4286d5010a1f35