[kotlin] Remove unnecessary wrapping to fix side-effect exception in Fleet
Hovering quickfixes in the list in Fleet shows a preview of the changes.
Quickfixes are actually run to capture these changes, but this doesn't bode
well with side-effects. With side-effects guards in place, Fleet usually
captures side-effect errors and stops the preview. However, due to
`ApplicationManager.getApplication().invokeLater` wrapping, this exception
capture cannot happen on Fleet side, and the exception escapes during
preview.

In this quickfix, the invokeLater wrapping doesn't actually seem necessary.
The `invoke` function of `KotlinQuickFixAction` already is run within
write action, and the changes are just project file updates (they are fast).

GitOrigin-RevId: 62e02fb41ca7f5dba09525881d790b44ceef302a