IJPL-60317 Add Expand node with single click setting
Use add/removeNotify in Tree to subscribe/unsubscribe to UI settings.
Only actually change the click count if the client has never changed
it directly, otherwise we're going to get very confusing behavior
(you call a function, it works for some time and then stops working).
If we decide to apply this to more trees, we'll have to remove
those client calls instead of overriding them here.

GitOrigin-RevId: d3924775669ad82fae7ae6d0ed4afc2b951fc4c4