IJPL-48486 Make Abbreviate Qualified Package Names always visible
We currently don't support menu items appearing and disappearing
without closing the menu first. Since this Project View menu is now
a multi-select menu that stays open after a click, it leads to confusing
behavior: one menu item is selected, another one should now be
visible, but it only appears the next time the menu is shown.

For now, fix by simply making the other menu item always
visible (as long as the current Project View pane supports
that option). It's now simply disabled if it isn't applicable
in the current state.

GitOrigin-RevId: fd8b5301cec3b713660df336575a77a370883bb1