IJPL-149149 Fix path combobox minimum size in new file chooser
By default ComboBox makes its minimum size equal to its
preferred size, and the latter is determined by the largest
element in the drop-down. For the path history combobox
it means that if we have one very long path in the history,
then the entire combobox will become very large,
and there will be no way to make it smaller.

Fix by simply turning this feature off for this specific combobox.
This means that its minimum size will be independent of
the preferred size, and in practice it makes the entire panel
stick to its preferred size instead, which is hardcoded to 600x450
and that makes much more sense than having one huge dialog
with no way to even resize it manually.

GitOrigin-RevId: 19635a580941f58f55eed6bfb60218777b21555c