[kotlin] fix completion from builtins in common sourcest
The problem was in fact that common stdlib does not contain builtins. In this case, builtin declarations should be resolved to the bundled builtins.

Completion works via `KtScopeProviderMixIn.getScopeContextForPosition`.
Builtins should come via default star-importing scope.
To collect callable and classifier names there, `IdeKotlinDeclarationProvider.getTopLevel(KotlinClassLikeDeclaration/Callables)NamesInPackage` are used which use the scope provided by `IdeKotlinByModulesResolutionScopeProvider`.

An additional change in `IdeKotlinByModulesResolutionScopeProvider` is needed to ensure that for the declarations from builtins we get `KtBuiltinsModule` instead of `KtNotUnderContentRootModule`

Tests were added in the previous commits as a part of KTIJ-28550

^KTIJ-29826 fixed


IJ-MR-138469

GitOrigin-RevId: 9b25f3f933e1299eea0a8dae3b6f15be37504a87