[terminal] IJPL-101589 Make options and arguments lazy fields instead of generators
It simplifies the spec and lowers the generators' cache pollution.
For now, there are no cases when ShellRuntimeContext is required to calculate the options or arguments. So, options and arguments lifetime can be the same as the lifetime of the command spec.
It can be changed back if we find a case when options or arguments need to be calculated depending on the shell state.

GitOrigin-RevId: 70f956bd9861fe95f51b7b2658e974e9a75ed972