[intention-preview] IntentionPreviewComputable: avoid no-side-effect check in ModCommand preview
In general, we assume that ModCommand-compatible actions behave well and don't produce any side effects. This differs from regular preview, which may automatically work for old non-migrated action. So it should be ok to lift the assertion. Sometimes, the assertion was triggered due to the fact that the first computation of code style settings fires change event for listeners in EDT. While it doesn't look pure, we may assume that it's a safe side effect.
Fixes IDEA-353083 SideEffectNotAllowedException for LaterInvocator.invokeLater during ModCommand.psiUpdate
Fixes EA-1183406 - SEGSENAE: IntentionPreviewComputable.getModActionPreview

GitOrigin-RevId: 60ec81ce1542d27364f7719ad1cbd2e575d81e77