IJPL-1002 Don't re-expand tree nodes if they were cached
When a TreeState is restored, if there was a cached presentation,
then it's immediately applied, and everything is expanded
instantly. If the user collapses one of these expanded nodes,
then it would re-expand again when the TreeState's visitor
finally loads the real node. This feels weird and not necessary,
because once the tree is shown, it should be up to the user
to expand/collapse anything at will.

Fix it by not expanding anything during the "expand" stage
in TreeState if there was a cached presentation. Still visit
the nodes to load them, but don't do anything with them.

GitOrigin-RevId: 9f805466d021e36e406ff41d9d3b8a936ccdcf84