IJPL-1002 Filter out cached nodes when visiting AsyncTreeModel
Cached nodes are useless for visitors because they don't contain
actual data and are usually not expected by the visitor.
If the model is being visited with allowLoading=true, then
it's impossible that a cached node ends up being visited
because the real children will be loaded first. However,
if allowLoading is false, then the model just returns whatever
children are already there. Previously these were either the real
children or an empty list, but now it's possible that there's
a list of cached nodes instead, which is undesirable.

Fix by filtering that list first. Return only the loaded children.

GitOrigin-RevId: a50f7f543705b2e4eb4039d72284f5c0e384bf54