[build scripts] support module-level libraries when determining the main location for libraries copied to multiple places
A module-level library can be also manually put to multiple places in the distribution via 'BaseLayoutSpec.withModuleLibrary' calls. So to ensure that the proper location is specified in module-descriptors.jar for such libraries and to fix RuntimeModuleRepositoryChecker.checkIntegrityOfEmbeddedProduct, we need to consider module-level libraries as well. Also, a new heuristic is added: if some of the copies is put to a JAR under a directory named 'client' or 'frontend', we use it as the main location to ensure that unrelated JARs won't be included in the classpath for the frontend process.

As a result, 'client/cwm-guest.jar' is chosen as the resource root for 'jetbrains.fleet.rpc.server' library (instead of 'cwm-plugin.jar').

GitOrigin-RevId: da584f6413494d5e8520cfb3e0b661d5f697ad40